v0.6: standalone program
Add function to build an electrode grid returning the electrodes (with 
segmentation) and nodes.

Add function to calculate harmonic impedance at every frequency and 
node.

Add function to do a general simulation outputing U, IL and IT.

Add standalone program that runs on the terminal. As of now, it only 
calculates the harmonic impedance of a rectangular grid with parameters 
given interactively by the user.

TODO: run standalone program based on input files.